Waste_manager/waste-management-app/requirements.txt
Flask==2.0.3
tensorflow==2.6.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
opencv-python==4.5.3.20210927
jupyter==1.0.0
Pillow==8.4.0
requests==2.26.0